Update more_trick_and_knowledge.py
here is a slight but important thing to know, we have to alsways remember to declare the variables outside the while loop. This will ensure that the values of these variables are retained between iterations of the loop. (this will avoid the values to be returned to zero, each time when the loop is restarted)