Update more_trick_and_knowledge.py
here we have created a simple list making program. you can notice now that we've used  two functions which are .lower() .strip(). Now we have used it directly next to the input and that basically mean that (in this case) we want our input to be directly converted to lower and stripped (no spaces before and after, they are erased)