Update more_trick_and_knowledge_2.py
here is a bug update of our HANGMAN game that we are building. we've added many features like the condition to check if the word has been fully guessed or not,  also we've added the feature of keeping the reccords of the number of letters that has been used so far.