Update more_trick_and_knowledge_2.py
here is anoter slight but important update. now we are adding a counter to numerate the tasks in our ToDo list in order to know how many tasks we have so far in our todo list. we can use this counter to set some specific limits to the list (like, limiting the number of task to a specific number based on the type of to do list chosen). MORE DETAILS WILL BE PROVIDED LATER ON.