Update more_trick_and_knowledge_2.py
here we are using the reading method we learned, but here we are basiclly taking the list created before and read after it is alredy converted back into code (here, back into list) and is then re-assigned to myEvents so that, even after the end of the program, the content will already be saved in myEvents... so that it cnnot be erased. NOW THE 2D LIST IS ASSIGNED TO "mtEvents" and can later be prettyPrinted.