Update more_trick_and_knowledge_2.py
As we said, we are updating our program with new features that we are learning. today is day 52/100 of the Replit python coding journey. today we have first added the TRY/EXCEPT concept in order to avoid errors when the slice of code we are launching is not working yet.