Update more_trick_and_knowledge_2.py
here we have added two new functionalities. We added, both the items view feature to be able to see what is inside the list, we also added the remove item, in order to be able to remove an item from the list before it is submitted to the local text file to be kept locally.