Update More_tricks_and_knowledge_3.py
these words inside [] are the keys that are soring the password as a value, and in order to access them, we need to use the key name and the os library: os.environ[name of the key storing the value]

we get the password that we are going to compare with the one that we asked for to the system

now we are just checking the details before to give any access to the program.

NOTE: remember, this is inside the loop, so that we can specify thre trial number etc.