Update More_tricks_and_knowledge_3.py
here is a simple program that basically takes the values that are provided to it (these values must respect a certain condition for them to be valid). Then we print the average of all those valid values. This is just an introduction to how to provide multiple input to one parameter or variable of a function, By just adding * which transforms it into a list of integers or float numbers.