Update More_tricks_and_knowledge_3.py
this is The OOP version of our average function. Basically here we have created a function for every task to be done in order to get the average number. First we collect the valid numbers with the AddNum() function, Then we use the collected values in our second function called CalculateAv(). THEN we initiate our Object so that we can use it to access all the functions of the class. Then at the end we are printing the average by using the name of the object then the name of the calculator function CalculateAv().